Saving all graph and HoF graphs implemented, with analysis
Graphs are all saved throughout topology optimization in a minimal format. Analysis, such as tracing back what history of that graph is, can then be done. Graphs are saved as JSON files, with info about the component models, graph adjacency, and component parameters. As well, unique IDs are attached to each new graph for tracing back